+++ src/API.ts
@@ -5,6 +5,18 @@ import FileMetadata from "./FileMetadata";
 import GlobalConfiguration from "./GlobalConfiguration";
 import OmnichannelChatToken from "./OmnichannelChatToken";
 
+enum AmsApiOperation {
+    Create = "Create",
+    Upload = "Upload"
+}
+
+enum DocumentTypes {
+    CreateDocumentType = 'sharing/file',
+    UploadDocumentType = 'original',
+    CreateImageType = 'pish/image',
+    UploadImageType = 'imgpsh'
+}
+
 enum HeadersName {
     Accept = 'Accept',
     AcceptEncoding = 'Accept-Encoding',
@@ -35,7 +47,7 @@ interface AMSHeaders {
     [HeadersName.AcceptEncoding]?: string;
 }
 
-const validImageTypes = ['jpeg', 'jpg', 'gif', 'png', 'bmp', 'tiff', 'jfif', 'webp'];
+const defaultSupportedImagesMimeTypes = ["image/jpeg", "image/png", "image/gif"];
 
 const patchChatToken = (chatToken: OmnichannelChatToken) => {
     // Temporary
@@ -82,23 +94,36 @@ const skypeTokenAuth = async (chatToken: OmnichannelChatToken): Promise<Response
     }
 }
 
-const createObject = async (id: string, file: File, chatToken: OmnichannelChatToken): Promise<AMSCreateObjectResponse> => {
+const defineSupportedImagesMimeTypes = (supportedImagesMimeTypes?: string[]) => {
+    return (supportedImagesMimeTypes && supportedImagesMimeTypes.length > 0) ?
+        supportedImagesMimeTypes :
+        defaultSupportedImagesMimeTypes;
+}
+
+const defineTypeForOperation = (fileType: string, apiOperation: string, supportedImagesMimeTypes?: string[]) => {
+    const mimeTypes = defineSupportedImagesMimeTypes(supportedImagesMimeTypes);
+
+    if (mimeTypes.includes(fileType.toLowerCase())) {
+        return apiOperation === AmsApiOperation.Create ? DocumentTypes.CreateImageType : DocumentTypes.UploadImageType;
+    }
+    return apiOperation === AmsApiOperation.Create ? DocumentTypes.CreateDocumentType : DocumentTypes.UploadDocumentType;
+}
+
+const createObject = async (id: string, file: File, chatToken: OmnichannelChatToken, supportedImagesMimeTypes?: string[]): Promise<AMSCreateObjectResponse> => {
     GlobalConfiguration.debug && console.log(`[API][createObject]`);
 
     const permissions = {
         [id]: ['read']
     };
-
+    const typeObject = defineTypeForOperation(file.type, AmsApiOperation.Create, supportedImagesMimeTypes);
     const body = {
         filename: file.name,
         permissions,
-        type: file.type.includes('image')? 'pish/image': 'sharing/file'
+        type: typeObject
     };
-
     patchChatToken(chatToken);
 
     const url = `${chatToken.amsEndpoint || chatToken?.regionGTMS?.ams}/v1/objects`;
-
     const headers = {
         ...createDefaultHeaders(chatToken.token),
         [HeadersName.ContentType]: MIMEType.applicationJson
@@ -121,13 +146,13 @@ const createObject = async (id: string, file: File, chatToken: OmnichannelChatTo
     }
 }
 
-const uploadDocument = async (documentId: string, file: File | AMSFileInfo, chatToken: OmnichannelChatToken): Promise<FileMetadata> => {
+const uploadDocument = async (documentId: string, file: File | AMSFileInfo, chatToken: OmnichannelChatToken, supportedImagesMimeTypes?: string[]): Promise<FileMetadata> => {
     GlobalConfiguration.debug && console.log(`[API][uploadDocument]`);
 
     patchChatToken(chatToken);
 
-    const url = `${chatToken.amsEndpoint || chatToken?.regionGTMS?.ams}/v1/objects/${documentId}/content/${file.type.includes('image')? 'imgpsh': 'original'}`;
-
+    const typeObject = defineTypeForOperation(file.type, AmsApiOperation.Upload, supportedImagesMimeTypes);
+    const url = `${chatToken.amsEndpoint || chatToken?.regionGTMS?.ams}/v1/objects/${documentId}/content/${typeObject}`;
     const headers = {
         ...createDefaultHeaders(chatToken.token),
         [HeadersName.ContentType]: MIMEType.applicationFormUrlEncoded
@@ -136,9 +161,8 @@ const uploadDocument = async (documentId: string, file: File | AMSFileInfo, chat
     const request = {
         headers,
         method: 'PUT',
-        body: (file as any).data? (file as AMSFileInfo).data: file as File  // eslint-disable-line @typescript-eslint/no-explicit-any
+        body: (file as any).data ? (file as AMSFileInfo).data : file as File  // eslint-disable-line @typescript-eslint/no-explicit-any
     };
-
     try {
         await fetch(url, request as RequestInit);
         const fileMetadata = {
@@ -156,12 +180,12 @@ const uploadDocument = async (documentId: string, file: File | AMSFileInfo, chat
     }
 }
 
-const getViewStatus = async (fileMetadata: FileMetadata, chatToken: OmnichannelChatToken): Promise<AMSViewStatusResponse> => {
+const getViewStatus = async (fileMetadata: FileMetadata, chatToken: OmnichannelChatToken, supportedImagesMimeTypes?: string[]): Promise<AMSViewStatusResponse> => {
     GlobalConfiguration.debug && console.log(`[API][getViewStatus]`);
 
     patchChatToken(chatToken);
 
-    const url = `${chatToken.amsEndpoint || chatToken?.regionGTMS?.ams}/v1/objects/${fileMetadata.id}/views/${validImageTypes.includes(fileMetadata.type)? 'imgpsh_fullsize_anim': 'original'}/status`;
+    const url = `${chatToken.amsEndpoint || chatToken?.regionGTMS?.ams}/v1/objects/${fileMetadata.id}/views/${defineSupportedImagesMimeTypes(supportedImagesMimeTypes).includes(fileMetadata.type) ? 'imgpsh_fullsize_anim' : 'original'}/status`;
 
     const headers = createDefaultHeaders(chatToken.token);
 
@@ -174,7 +198,7 @@ const getViewStatus = async (fileMetadata: FileMetadata, chatToken: OmnichannelC
         const response = await fetch(url, request);
         const jsonResponse = await response.json();
 
-        const {content_state, view_state, view_location} = jsonResponse;
+        const { content_state, view_state, view_location } = jsonResponse;
 
         if (!view_location) {
             throw new Error('view_location is empty');
@@ -195,7 +219,7 @@ const getViewStatus = async (fileMetadata: FileMetadata, chatToken: OmnichannelC
     }
 }
 
-const getView = async (fileMetadata: FileMetadata, viewLocation: string, chatToken: OmnichannelChatToken): Promise<Blob> => {
+const getView = async (fileMetadata: FileMetadata, viewLocation: string, chatToken: OmnichannelChatToken, supportedImagesMimeTypes?: string[]): Promise<Blob> => {
     GlobalConfiguration.debug && console.log(`[API][getView]`);
 
     patchChatToken(chatToken);
@@ -204,7 +228,7 @@ const getView = async (fileMetadata: FileMetadata, viewLocation: string, chatTok
 
     const headers = createDefaultHeaders(chatToken.token);
 
-    if (validImageTypes.includes(fileMetadata.type)) {
+    if (defineSupportedImagesMimeTypes(supportedImagesMimeTypes).includes(fileMetadata.type)) {
         headers[HeadersName.Accept] = 'image/webp,image/ *,*/*;q=0.8';
         headers[HeadersName.AcceptEncoding] = 'gzip, deflate, sdch, br';
     }